+++ CHANGELOG.md
@@ -14,6 +14,7 @@ at anytime.
 
 ### Changed
   * Do not catch base exception in API command resolve
+  * Remove deprecated `lbrynet.metadata` and update what used it to instead use `lbryschema`
   *
 
 ### Fixed